test(cosmic-swingset): Introduce a swingset instantiation discontinuity to the run policy test
An unrelated CI test experienced a failure when budgeted vat cleanup had
already removed a strict subset of a slowly-terminating vat's kvStore
entries that included the one with key "${vatID}.o.nextID"...
`provideVatKeeper` uses such keys as a "vat is alive" predicate that
must pass before it spins up any new keeper (whose primary purpose is
providing an interface for a *live* vat).

This is not an issue within a single swingset instantiation, in which
`provideVatKeeper` is memoized. But that cache starts empty with each
process, and the attempt to instantiate a new keeper for a slowly-
terminating vat in this state fails.

run-policy.test.js now covers this scenario.