chore(smart-wallet): less noisy watch outcome (#10515)
closes: #10513

## Description
Use `when` instead of `asPromise` in the smart-wallet offer outcome logic. Reduce noise on upgrades for disconnected promises which are not actually used for production logic.

### Security Considerations

None

### Scaling Considerations

Removes some unnecessary watch bookkepping.

### Documentation Considerations
None. Added a comment for maintainers explaining the choice of `when`.

### Testing Considerations
Expecting existing tests to keep passing. The only functional difference is the removal of an unnecessary warning diagnostic.

### Upgrade Considerations

Requires an upgrade of the walletFactory vat.