chore(.github): Add release verification prompts to PR template section "Upgrade Considerations" (#10444)
## Description
A topic in last week's kernel meeting was the burden of verifying releases, and even of determining _what_ to verify and _how_ to do so. This is an attempt to make that explicit at a much earlier point in development, with hints regarding future automation potential but no firm commitment of such. Assuming this lands, I expect release instructions to include collecting steps from all constituent PRs and post-release cleanup of any relevant scripts.

### Security Considerations
n/a

### Scaling Considerations
n/a

### Documentation Considerations
n/a

### Testing Considerations
n/a

### Upgrade Considerations
This is intended to increase the maturity of our upgrade process.