Add AssetReserve to Upgrade 19 with A3P test coverage (#10541)
closes: https://github.com/Agoric/agoric-sdk/issues/10399

## Description
Adds the upgrade of the reserve to upgrade 19. To ensure that this upgrade will succeed, we've also added test coverage in A3P. One caveat we discovered while testing was that the `adminFacet` that's produced with the `reserveKit` was referencing the `adminFacet` of the governor. Luckily, we were able to get the `adminFacet` of the reserve contract by calling `getAdminFacet` on the governor's `creatorFacet`.

### Upgrade Considerations

This PR adds the reserve to the list of vat upgrades for upgrade-19. These upgrades are currently commented out as we are still in the process of getting upgrade-18 out the door. We've added A3P tests to ensure that the upgrade is successful and that assets are still in the reserve after the upgrade.