Add PSM upgrade for upgrade-19 (#10730)
closes: https://github.com/Agoric/agoric-sdk/issues/10403

## Description
Adds the upgrade of the psm contracts to upgrade 19. To ensure that this upgrade will succeed, we've also added test coverage in A3P.

### Upgrade Considerations

This PR adds the PSM contract to the list of vat upgrades. The PSM upgrade proposal will upgrade all PSM instances on the chain. We've added A3P tests to ensure that the upgrade is successful and that assets are still in the reserve after the upgrade.