chore: update  installations when restarting or upgrading (#10999)
## Description

I've known for a while that there was something magic about the `installations` keyword in `getManifestForFoo` in coreEval proposals, but I was never able to track it down. Now I've found the relevant code. In [`coreProposalBehavior`](https://github.com/Agoric/agoric-sdk/blob/a67bf2862935b4187f423813cfa327c0fea187b8/packages/deploy-script-support/src/coreProposalBehavior.js#L169) (which is copied into generated proposals), if the manifest contains named `installations`, then they are added to **both agoricNames and the bootstrap promiseSpace**.  When upgrading a contract to a new installation, the installation should nearly always be updated this way.

This PR updates all the proposals/scripts currently slated for upgrade-19 to include their upgraded bundles in `installations` so future references will automatically pick them up.

### Security Considerations

related to security and reliability of the contracts, not the chain.

### Scaling Considerations

not relevant.

### Documentation Considerations

I updated the *README* in deploy-script-support.

### Testing Considerations

The `n:upgrade-next` and `z:acceptance` tests continue to pass. 

### Upgrade Considerations

PSM is the only contract being upgraded in Upgrade19 that starts up multiple copies, and it was already doing this correctly.