fix(agoric-cli): agops oracle pushPriceRound will fail with some prices (#11009)
closes: #10122
refs: #XXXX

## Description
Due to floating point precision limitation, some numbers (like `8.3`) when multiplied with the scaler `1000000` can result into numbers not convertable into `BigInt`. This created problems when generating the `PushPrice` wallet action for the oracles (as mentioned in the issue). A similar issue could arise when using a price of more than 6 decimal points. This PR fixes that by rounding off the price before the `BigInt` conversion.
Also update the `makeOracleCommand` to use powers instead of global variables
Build success for the added test:
https://github.com/Agoric/agoric-sdk/actions/runs/13356850486/job/37301216448?pr=11009#step:5:94

### Security Considerations
None

### Scaling Considerations
None

### Documentation Considerations
None

### Testing Considerations
Test added in the PR

### Upgrade Considerations
None