fix(swing-store): Update archiveTranscript to wait for write stream conclusion (#11294)
closes: #11293

## Description
`archiveTranscript` creates a WriteStream `writer` and a gzip stream `gzip`, connects them by `gzip.pipe(writer)`, and then waits for `streamFinished(gzip)` before returning. But `writer` can still have pending activity even after `gzip` finished, which seems to be the root cause of #11293.

This PR updates `archiveTranscript` to instead wait for conclusion of _`writer`_, currently by listening for error/finish/close events (but I'm open to alternatives that also accomplish the same goal).

### Security Considerations
n/a

### Scaling Considerations
n/a

### Documentation Considerations
n/a

### Testing Considerations
This is fixing racy behavior, and thus impractical to test specifically. We'll instead rely on the existing coverage that alerted us to begin with.

### Upgrade Considerations
n/a