feat(base-zone): attenuator as simplified revocable (#11314)
closes: https://github.com/Agoric/agoric-sdk/issues/11309
refs: https://github.com/Agoric/agoric-sdk/pull/11307

## Description

This attenuator implementation is just a simplification of the revocable kit implementation in prepare-revocable.js. The revocable kit provides both attenuation and revocation, since we can have both for the price of just one level of indirection. But if you don't need revocation, the revocable kit has more API complexity than you need.

Because the attenuator just uses an exo class rather than a class kit, it cannot support the privilege separation of a distinct revoker facet. But it can still support `selfRevoke` as a separate `extraMethod`, as shown in a testcase at `prepare-attenuator.test.js`.

- [x] Understand the `attenuate` in `@agoric/internal`, and, if possible, reconcile them. Or deprecate one in favor of the other.
  - That `attenuate` is quite different. It is more of a deep pick appropriate for own data properties, or very specifically for own non-this-sensitive methods. The attenuator it creates necessarily has own properties for these deeply picked properties. Whereas, this PR creates an exo attenuator with the exposed properties on its own exo class. It is built assuming that the underlying is an exo, but it would probably work for a far object as well. The key is it makes no assumption about where in the underlying's inheritance chain the method is found. The attenuator in this PR has no support for data properties.
- [x] Similarly, understand the `prepareAttenuator` in `@agoric/internal`.
  - That `prepareAttenuator` is similar in many ways to the attenuator in this PR. However, it depends on a `Callback` concept that IMO makes it significantly harder to understand. What extra expressiveness to we get for that extra complexity? Do we use any of that extra expressiveness? I don't know. To be investigated after this PR. Attn @michaelfig 

### Security Considerations

Attenuation helps POLA, which helps security.

### Scaling Considerations

As with revocation, the cost of attenuation is one more level of indirection. So, for most purposes, none.

In particular, a concern was raised about attenuators having per-instance methods as opposed to per-class methods. As with revocation, this implementation uses only per-class methods.

### Documentation Considerations

Both revocation and attenuation need to be documented.

### Testing Considerations

Attenuation testing included, based on the existing revocation tests.

### Upgrade Considerations

Done as prepare patterns, so should be independently upgradable. Including change to what methods are included and what `extraMethods` are added.

Note that when `@agoric/base-zone` moves to endo, this attenuator should move with it. The meaning is clearly as useful for non-agoric endo users. Attn @kriskowal