refactor: use endo `getMethodNames` rather than agoric-sdk placeholder (#11327)
closes: #XXXX
refs: #XXXX

## Description

We only did the first half of migrating `getMethodNames` from `@agoric/internal` to `@endo/eventual-send`, which is that we copied it there, waiting to delete the original and change dependents after an agoric-endo sync which happened a long time ago.

This PR takes those other two steps. 
- It replaces the agoric-sdk implementation of `getMethodNames` with one that trivially forwards to the one in endo, and deprecates this trivial forwarding function.
- It changes agoric-sdk importers of the one in agoric-sdk to import from endo instead.

Types aside, these two implementations were exact copies. But in trying to reconcile the types I ran into a bit of trouble. In agoric-sdk `combineInvitationMakers` depended on the method names being strings, which was somehow supported by the types of the `getMethodNames` in agoric-sdk but not by the less precise types of the one in endo. Fortunately there was a non-redundant `getStringMethodNames` that filters for strings only, so I changed `combineInvitationMakers` to use that instead.
- [ ] in a separate PR, revise the `getMethodNames` in endo to use the more precise types, solving the resulting type conflicts that causes in endo.

### Security Considerations

AFAICT, `combineInvitationMakers`, if given an object with symbol-named methods, would throw an unintended error where it assumes the results are only strings `n.startsWith('__')`. I say "unintended" because, to my eyes, this code was not written to be defensive against that case. Rather it assumes that case won't happen, so it happens to error when that assumption is violated.

But `getStringMethodNames` filters out symbol-names, resulting in `combineInvitationMakers` ignoring these, rather than throwing an unintended error. This in theory could weaken security, but I think it is highly unlikely. So unlikely that I label this PR as a refactor.

### Scaling Considerations
none
### Documentation Considerations
one less thing to document.
### Testing Considerations
As a pure refactor, all we normally ask is that no tests needed to be revised, aside from the refactor itself, and that all tests still pass. They do.
### Upgrade Considerations
none