fix(cosmic-proto): make `address-hooks.js` more tolerant
We need the `golang/cosmos` semantics to be able to handle malformed
bech32 addresses, and definitely addresses that weren't address
hooks.  Change the behaviour of `decodeAddressHook` to return a
baseAddress matching the input (with no query parameters) instead of
throwing an error.

If the old behaviour is desired, use `mustDecodeAddressHook`.